requirements.txt
# --- CORE FRAMEWORK ---
fastapi
uvicorn
jinja2
python-multipart
requests
sqlmodel
hailo_platform
psycopg2-binary
reportlab

# --- IMAGE PROCESSING ---
pillow
numpy
piexif
opencv-python-headless  # <--- CRITICAL ADDITION FOR DOCKER

# --- AI & ANALYTICS ---
nudenet                 # The CPU Fallback Brain
onnxruntime             # The engine for NudeNet

# --- REPORTING ---
reportlab==4.0.7        # For generating Client PDF Reports
python-dateutil         # Robust date parsing
